requirements.txt
streamlit
langchain
sentence-transformers
faiss-cpu
pypdf
huggingface_hub 